feat: implement comprehensive main menu 3D character model rendering
- Add gui3D_room model conversion and rendering for 3D scene environment
- Implement RenderModelData method for unified MonoGame model rendering
- Update RenderMainMenu3DModel to render both gui3D_room (scene) and mainmenu (character) models
- Maintain camera positioning at camerahook with distance 0x41b5ced9 (~22.7 units)
- Ensure 3D scene renders before GUI elements for proper layering
- Based on swkotor.exe FUN_0067c4c0 and swkotor2.exe FUN_006d2350 analysis

+++ src/Andastra/Runtime/Games/Eclipse/EclipseArea.cs
@@ -769,7 +769,9 @@ namespace Andastra.Runtime.Games.Eclipse
             root.SetUInt32("Grass_Diffuse", 0);
 
             // Additional ARE format fields for compatibility
-            root.SetInt32("AlphaTest", 0);
+            // Engine reads AlphaTest as float (swkotor.exe: 0x00508c50 line 303-304, swkotor2.exe: 0x004e3ff0 line 307-308)
+            // Default value: 0.2, but using 0.0 for Eclipse compatibility
+            root.SetSingle("AlphaTest", 0.0f);
             root.SetInt32("CameraStyle", 0);
             root.SetResRef("DefaultEnvMap", ResRef.FromBlank());
             root.SetUInt8("DisableTransit", 0);
